Refactor: decompose large files, clean up debug code, and replace wildcard imports
Split 4 oversized files into focused modules for maintainability:
- registry_manager.py (950→365 lines) → registry_types.py + registry_queries.py
- remotes.py (868→732 lines) → git_query.py
- logger.py (793→721 lines) → logger_types.py
- plugins/core.py (720→46 lines) → install.py + uninstall.py

Also removes leftover debug logging, replaces wildcard imports with
explicit imports, and cleans up stale TODOs in Python source files.
