+++ src/gurk/__init__.py
@@ -2,7 +2,28 @@
 from importlib.metadata import version
 
 from .lib.core.context import Logger, LoggerSeverity
-from .lib.helpers import *
-from .lib.utils import *
+from .lib.helpers import (
+    BuiltinInstallCommands,
+    InstallCommandsBase,
+    add_alias,
+    get_clean_lines,
+    install_packages_from_list,
+    install_packages_from_txt_file,
+    parse_task_args,
+)
+from .lib.utils import (
+    PatternCollection,
+    dump_toml,
+    dump_yaml,
+    extract_url,
+    git_clone,
+    is_git_repo,
+    is_url,
+    load_toml,
+    load_yaml,
+    resolve_package_path,
+    revert_sudo_permissions,
+    run_script_function,
+)
 
 __version__ = version("gurk")
