+++ src/gurk/lib/core/context/__init__.py
@@ -1,7 +1,7 @@
 from .gurk_context import GurkContext
 from .logger import Logger, LoggerSeverity, TaskTerminationType, get_logger
-from .registry_manager import (
-    get_plugin_directories,
+from .registry_manager import get_plugin_directories
+from .registry_queries import (
     get_registries,
     get_registry_files,
     is_plugin_registered,
